requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework==3.11.1
markdown==3.2.2 
django-filter==2.3.0
requests==2.24.0 